Changes to scanning config
Needed to align with latest spec proposal

Existing scanning code would not reliably find the longest match when a
class or package name matches multiple entries from the same config
variable.

E.g. if SCAN_PACKAGES was "com.example.a, com.example.a.b.c", then when
matching "com.example.a.b.c", it might report the longest match as being
just "com.example.a".

Change to matching against each entry in each config variable
individually, rather than creating a single regex which would match any
of them.

Also, if the config is interpreted as a regex, match it against a class
or package name directly rather than requiring the match to also be a
prefix or suffix.