TCK runner: Update the archive using BeforeDeploy
Previously an ApplicationArchiveProcessor was used to run the
implementation and generate the OpenAPI document in the archive before
deploying it. However, and ApplicationArchiveProcessor is only called
for _testable_ archives (archives where tests run on the server),
whereas to pass the TCK we need to process all the archives.

Instead, use the BeforeDeploy event to update the archive which is
called for every deployment.