Gradle plugin: Fix circular task dependency with Quarkus plugin (#1344)
The Smallrye Gradle plugin adds the generated OpenAPI schema files (YAML + JSON) to the generated jar file. This was achieved using via the `processResources` task, which unfortunately is not "compatible" with the Quarkus plugin (see below output). The fix is to add the generated schema files directly to the `jar` task.

Tests have been added to verify interopability with the Quarkus plugin, also reproducers for #1335 (without the production code changes of course):
```
Circular dependency between the following tasks:
:compileJava
+--- :quarkusGenerateCode
|    \--- :processResources
|         \--- :generateOpenApiSpec
|              +--- :compileJava (*)
|              +--- :quarkusGenerateCode (*)
|              \--- :quarkusGenerateCodeDev
|                   \--- :processResources (*)
\--- :quarkusGenerateCodeDev (*)
```

Fixes #1335