Support generation of OpenAPI v3.0
Support for OpenAPI v3.0 is implemented as a transformation on input and
output. The changes between the two formats are minimal and mostly in
the Schema object.

When reading or writing the OpenAPI version is read from the document
or model respectively.

When generating a document for an application, config is read and if a
verison is configured the OpenAPI version in the model will be updated
before it is output.