Fix OpenAPI 3.0 output of nullable references
When a nullable field or parameter references a non-nullable schema, we
generate a schema for the field which uses anyOf to permit null. Any
informational properties or additional assertions defined on the field
are also put on the field schema and the type property is also retained.

When we were transforming this field schema for OpenAPI 3.0, we were
only doing certain transformations if the type property is not present,
which doesn't handle the case above.

Ensure the transformation to and from OpenAPI 3.0 schemas handles these
nullable references which include the type property.